Add hdf and mkl libraries to intel.mk (#6)
* Change URL links to github relative links to files

Using relative links will allow the links to point to files on different branches

* Pushing build info

* Update and rename README.AM4_build to AM4_build.README

* Rename AM4_build.README to README_AM4_build.md

* Update README.md

* Update README.md

* Correct errors in the compile script

The compile script had a few errors that kept it from generating the Makefiles.

Although, I think there are still issues that may keep the Makefiles from working correctly.

* Add Makefiles for components

Following the SM2 public release, this adds pregenerated Makefiles that remove the need for external users to use mkmf to generate Makefiles.

* adding runscript from T.Robinson

* Update README.md

* Update README.md

* Updated CPP macros to allow compile to complete

* Update Makefile to not pass OPENMP to ocean

MOM6, current, does not support openMP.  The Makefile now does not pass the OPENMP macro to the MOM6 submake.

* Add missing -fp-model source option to make template

The -fp-model source option is needed to allow the model to run.  Also corrected the NO_OVERRIDE_LIMIT macro option

* Changed total_npes to match settings in the sample input files

The data files linked to in the documentation use 216 MPI processes for the run.  Updated the run script to match.

* Update README.md

Includes partial build and run instructions.
Include updated links to data files.

* README add links to submodule repositories

Add links to the components added as submodule in the README.md file.

* Correct spelling in README.md

Correct spelling mistakes in README.md

* Remove build directory

Removed to no longer required build directory.

* Add git hashes for GFDL internal git repos

Added the git commit has for the components currently stored in the GFDL intenral git repositories.  This is to help if/when there are support or community supplied bug fixes.

Also corrected the clone --recursive option in the cloning instructions.

* Fix intel.mk template to include hdf and mkl libraries.

Also fix a few errors in the run script sample.

* Set omp_threads in run script to match input.nml in AM4 input archive

Corrected spelling of "using" in comment